discord-gallerydl/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
decorator==5.1.1
discord.py==1.7.3
gallery-dl==1.20.3
idna==3.3
multidict==6.0.2
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
validators==0.18.2
yarl==1.7.2
